refactor(421): extract runtime-trap machinery to vera/runtime/traps.py
Step 2 of the codegen/api.py decomposition (#421): the runtime-trap classification + source-backtrace machinery (WasmTrapError, TrapFrame, _find_frames_in_exception_chain, _resolve_trap_frames, _TRAP_FIX_PARAGRAPHS, _classify_trap) moves out of api.py into a new vera/runtime/ package (traps.py) -- the first member of the execution-runtime package. The block is self-contained (no api.py-internal deps); wasmtime is annotation-only there so it sits under TYPE_CHECKING.

api.py re-exports WasmTrapError (part of execute()'s public contract, kept importable for cli.py + the #734 harness) and imports _classify_trap/_resolve_trap_frames internally; the Diagnostic annotation import shifts to api.py's TYPE_CHECKING block. test_runtime_traps.py imports the helpers from runtime.traps. api.py 4,247->3,790 lines. mypy + ruff clean; the trap suite + #734 gate (88) stay green.
