+++ CHANGELOG.md
@@ -14,7 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 
-- **`codegen/api.py` decomposition — step 1: ADT memory layout ([#421](https://github.com/aallan/vera/issues/421))**.  Lifted the pure ADT memory-layout utilities (`ConstructorLayout`, `_validate_wrap_handle`, `_wasm_type_size` / `_wasm_type_align` / `_align_up`) out of the 4,358-line `vera/codegen/api.py` into a focused `vera/codegen/memory.py`, with the codegen consumers importing them from there directly.  The public surface is unchanged — `from vera.codegen import ConstructorLayout` and `from vera.codegen.api import compile` / `execute` / `CompileResult` / `ExecuteResult` all still resolve — and the #734 characterization harness stays green.  First of a per-concern series that lifts the runtime out of the public-API module so the deferred whole-`vera/` mutation sweep ([#387](https://github.com/aallan/vera/issues/387)) becomes tractable.  Internal refactor; no behaviour change.
+- **`codegen/api.py` decomposition ([#421](https://github.com/aallan/vera/issues/421))**.  Splitting the 4,358-line public-API module into focused units, one layer per commit behind the #734 characterization harness.  So far: (1) the pure ADT memory-layout utilities (`ConstructorLayout`, `_validate_wrap_handle`, `_wasm_type_size` / `_wasm_type_align` / `_align_up`) → a focused `vera/codegen/memory.py`; (2) runtime-trap classification and source-backtrace resolution (`WasmTrapError`, `TrapFrame`, `_classify_trap`, `_resolve_trap_frames`) → a new `vera/runtime/` package (`traps.py`).  The public surface is unchanged — `from vera.codegen import ConstructorLayout` and `from vera.codegen.api import compile` / `execute` / `CompileResult` / `ExecuteResult` / `WasmTrapError` all still resolve.  api.py is down from 4,358 to 3,790 lines so far; the wasmtime host-binding families follow.  Lifting the runtime out of the public-API module also unblocks the deferred whole-`vera/` mutation sweep ([#387](https://github.com/aallan/vera/issues/387)).  Internal refactor; no behaviour change.
 
 ## [0.0.177] - 2026-06-21
