+++ KNOWN_ISSUES.md
@@ -66,7 +66,7 @@ Files that have grown beyond a comfortable size and need decomposition. None of 
 |------|-------|-------------|-------|
 | `tests/test_codegen.py` | 19,570 | Split into feature-focused test files (literals, arithmetic, control flow, strings, arrays, collections, effects, data types). The file has nearly doubled since the issue was filed; roadmap Tier 3. | [#419](https://github.com/aallan/vera/issues/419) |
 | `tests/test_checker.py` | 5,939 | Split into phase-focused test files (types, functions, effects, contracts, modules, errors). Currently a not-doing-now item — a close-or-annotate decision is pending. | [#420](https://github.com/aallan/vera/issues/420) |
-| `vera/codegen/api.py` | 4,253 | Extract the wasmtime execution layer into a `vera/runtime/` package, one host-binding family per PR. The characterization harness (#734) and the GC bucket-as-truth migration (#706) land first; the file has doubled since the issue was filed. | [#421](https://github.com/aallan/vera/issues/421) |
+| `vera/codegen/api.py` | 3,790 | Extract the wasmtime execution layer into a `vera/runtime/` package, layered by concern behind the #734 characterization harness — memory-layout utils → `codegen/memory.py` and trap machinery → `runtime/traps.py` extracted so far; the host-binding families follow. | [#421](https://github.com/aallan/vera/issues/421) |
 
 ## Test coverage gaps
