+++ vera/runtime/__init__.py
@@ -1,0 +1,6 @@
+"""Vera execution runtime.
+
+The wasmtime host layer, extracted from `vera/codegen/api.py` (#421).
+Trap classification and source-backtrace resolution today; the
+host-binding families follow in later layers.
+"""
